+++ include/reactphysics3d/containers/Stack.h
@@ -79,7 +79,7 @@ class Stack {
 
             mArray = newArray;
 
-            mCapacity = capacity;
+            mCapacity = (uint)capacity;
         }
 
     public:
