+++ src/engine/EntityManager.cpp
@@ -55,7 +55,7 @@ Entity EntityManager::createEntity() {
         // Create a new indice
         index = static_cast<uint32>(mGenerations.size()) - 1;
 
-        assert(index < (1 << Entity::ENTITY_INDEX_BITS));
+        assert(index < (uint32_t)(1 << Entity::ENTITY_INDEX_BITS));
     }
 
     // Return a new entity
